requirements.txt
gspread==6.1.4
google-auth==2.38.0
chromadb==0.5.23
openai==1.61.0
streamlit==1.41.1
tiktoken==0.8.0
tenacity==9.0.0
python-dotenv==1.0.1
